Move the image example to examples/oci-image-verification
