+++ cmd/sigstore-go/main.go
@@ -15,11 +15,9 @@
 package main
 
 import (
-	"bytes"
 	"crypto"
 	"crypto/ecdsa"
 	"crypto/x509"
-	"encoding/base64"
 	"encoding/hex"
 	"encoding/json"
 	"encoding/pem"
@@ -28,14 +26,6 @@ import (
 	"fmt"
 	"os"
 	"time"
-
-	"github.com/google/go-containerregistry/pkg/crane"
-	"github.com/google/go-containerregistry/pkg/name"
-	v1 "github.com/google/go-containerregistry/pkg/v1"
-	"github.com/google/go-containerregistry/pkg/v1/remote"
-	protobundle "github.com/sigstore/protobuf-specs/gen/pb-go/bundle/v1"
-	protocommon "github.com/sigstore/protobuf-specs/gen/pb-go/common/v1"
-	protorekor "github.com/sigstore/protobuf-specs/gen/pb-go/rekor/v1"
 
 	"github.com/sigstore/sigstore-go/pkg/bundle"
 	"github.com/sigstore/sigstore-go/pkg/root"
@@ -58,7 +48,6 @@ var trustedPublicKey *string
 var trustedrootJSONpath *string
 var tufRootURL *string
 var tufDirectory *string
-var ociImage *string
 
 func init() {
 	artifact = flag.String("artifact", "", "Path to artifact to verify")
@@ -75,9 +64,8 @@ func init() {
 	trustedrootJSONpath = flag.String("trustedrootJSONpath", "examples/trusted-root-public-good.json", "Path to trustedroot JSON file")
 	tufRootURL = flag.String("tufRootURL", "", "URL of TUF root containing trusted root JSON file")
 	tufDirectory = flag.String("tufDirectory", "tufdata", "Directory to store TUF metadata")
-	ociImage = flag.String("ociImage", "", "OCI image to verify")
 	flag.Parse()
-	if flag.NArg() == 0 && *ociImage == "" {
+	if flag.NArg() == 0 {
 		usage()
 		os.Exit(1)
 	}
@@ -96,19 +84,11 @@ func main() {
 }
 
 func run() error {
-	var b *bundle.ProtobufBundle
-	var err error
-
-	if *ociImage != "" {
-		// Build a bundle from OCI image reference and get its digest
-		b, artifactDigest, err = bundleFromOCIImage(*ociImage)
-	} else {
-		// Load the bundle from file
-		b, err = bundle.LoadJSONFromPath(flag.Arg(0))
-	}
+	b, err := bundle.LoadJSONFromPath(flag.Arg(0))
 	if err != nil {
 		return err
 	}
+
 	if *minBundleVersion != "" {
 		if !b.MinVersion(*minBundleVersion) {
 			return fmt.Errorf("bundle is not of minimum version %s", *minBundleVersion)
@@ -233,215 +213,4 @@ func trustedPublicKeyMaterial(pk crypto.PublicKey) *root.TrustedPublicKeyMateria
 		}
 		return &nonExpiringVerifier{verifier}, nil
 	})
-}
-
-// bundleFromOCIImage returns a ProtobufBundle based on OCI image reference.
-func bundleFromOCIImage(imageRef string) (*bundle.ProtobufBundle, *string, error) {
-	// 1. Get the simple signing layer
-	simpleSigning, err := simpleSigningLayerFromOCIImage(imageRef)
-	if err != nil {
-		return nil, nil, fmt.Errorf("error getting simple signing layer: %w", err)
-	}
-	// 2. Build the verification material for the bundle
-	verificationMaterial, err := getBundleVerificationMaterial(simpleSigning)
-	if err != nil {
-		return nil, nil, fmt.Errorf("error getting verification material: %w", err)
-	}
-	// 3. Build the message signature for the bundle
-	msgSignature, err := getBundleMsgSignature(simpleSigning)
-	if err != nil {
-		return nil, nil, fmt.Errorf("error getting message signature: %w", err)
-	}
-	// 4. Construct and verify the bundle
-	pb := protobundle.Bundle{
-		MediaType:            bundle.SigstoreBundleMediaType01,
-		VerificationMaterial: verificationMaterial,
-		Content:              msgSignature,
-	}
-	bun, err := bundle.NewProtobufBundle(&pb)
-	if err != nil {
-		return nil, nil, fmt.Errorf("error creating bundle: %w", err)
-	}
-	// 5. Return the bundle and the digest of the simple signing layer (this is what is signed)
-	return bun, &simpleSigning.Digest.Hex, nil
-}
-
-// simpleSigningLayerFromOCIImage returns the simple signing layer from the OCI image reference
-func simpleSigningLayerFromOCIImage(imageRef string) (*v1.Descriptor, error) {
-	// 1. Get the image reference
-	ref, err := name.ParseReference(imageRef)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing image reference: %w", err)
-	}
-	// 2. Get the image descriptor
-	desc, err := remote.Get(ref)
-	if err != nil {
-		return nil, fmt.Errorf("error getting image descriptor: %w", err)
-	}
-	// 3. Get the digest
-	digest := ref.Context().Digest(desc.Digest.String())
-	h, err := v1.NewHash(digest.Identifier())
-	if err != nil {
-		return nil, fmt.Errorf("error getting hash: %w", err)
-	}
-	// 4. Construct the signature reference - sha256-<hash>.sig
-	sigTag := digest.Context().Tag(fmt.Sprint(h.Algorithm, "-", h.Hex, ".sig"))
-	// 5. Get the manifest of the signature
-	mf, err := crane.Manifest(sigTag.Name())
-	if err != nil {
-		return nil, fmt.Errorf("error getting signature manifest: %w", err)
-	}
-	sigManifest, err := v1.ParseManifest(bytes.NewReader(mf))
-	if err != nil {
-		return nil, fmt.Errorf("error parsing signature manifest: %w", err)
-	}
-	// 6. Ensure there is at least one layer and it is a simple signing layer
-	if len(sigManifest.Layers) == 0 || sigManifest.Layers[0].MediaType != "application/vnd.dev.cosign.simplesigning.v1+json" {
-		return nil, fmt.Errorf("no suitable layers found in signature manifest")
-	}
-	// 7. Return the layer - most probably there are more layers (one for each signature) but verifying one is enough
-	return &sigManifest.Layers[0], nil
-}
-
-// getBundleVerificationMaterial returns the bundle verification material from the simple signing layer
-func getBundleVerificationMaterial(manifestLayer *v1.Descriptor) (*protobundle.VerificationMaterial, error) {
-	// 1. Get the signing certificate chain
-	signingCert, err := getVerificationMaterialX509CertificateChain(manifestLayer)
-	if err != nil {
-		return nil, fmt.Errorf("error getting signing certificate: %w", err)
-	}
-	// 2. Get the transparency log entries
-	tlogEntries, err := getVerificationMaterialTlogEntries(manifestLayer)
-	if err != nil {
-		return nil, fmt.Errorf("error getting tlog entries: %w", err)
-	}
-	// 3. Construct the verification material
-	return &protobundle.VerificationMaterial{
-		Content:                   signingCert,
-		TlogEntries:               tlogEntries,
-		TimestampVerificationData: nil,
-	}, nil
-}
-
-// getVerificationMaterialTlogEntries returns the verification material transparency log entries from the simple signing layer
-func getVerificationMaterialTlogEntries(manifestLayer *v1.Descriptor) ([]*protorekor.TransparencyLogEntry, error) {
-	// 1. Get the bundle annotation
-	bun := manifestLayer.Annotations["dev.sigstore.cosign/bundle"]
-	var jsonData map[string]interface{}
-	err := json.Unmarshal([]byte(bun), &jsonData)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshaling json: %w", err)
-	}
-	// 2. Get the log index, log ID, integrated time, signed entry timestamp and body
-	logIndex, ok := jsonData["Payload"].(map[string]interface{})["logIndex"].(float64)
-	if !ok {
-		return nil, fmt.Errorf("error getting logIndex")
-	}
-	li, ok := jsonData["Payload"].(map[string]interface{})["logID"].(string)
-	if !ok {
-		return nil, fmt.Errorf("error getting logID")
-	}
-	logID, err := hex.DecodeString(li)
-	if err != nil {
-		return nil, fmt.Errorf("error decoding logID: %w", err)
-	}
-	integratedTime, ok := jsonData["Payload"].(map[string]interface{})["integratedTime"].(float64)
-	if !ok {
-		return nil, fmt.Errorf("error getting integratedTime")
-	}
-	set, ok := jsonData["SignedEntryTimestamp"].(string)
-	if !ok {
-		return nil, fmt.Errorf("error getting SignedEntryTimestamp")
-	}
-	signedEntryTimestamp, err := base64.StdEncoding.DecodeString(set)
-	if err != nil {
-		return nil, fmt.Errorf("error decoding signedEntryTimestamp: %w", err)
-	}
-	// 3. Unmarshal the body and extract the rekor KindVersion details
-	body, ok := jsonData["Payload"].(map[string]interface{})["body"].(string)
-	if !ok {
-		return nil, fmt.Errorf("error getting body")
-	}
-	bodyBytes, err := base64.StdEncoding.DecodeString(body)
-	if err != nil {
-		return nil, fmt.Errorf("error decoding body: %w", err)
-	}
-	err = json.Unmarshal(bodyBytes, &jsonData)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshaling json: %w", err)
-	}
-	apiVersion := jsonData["apiVersion"].(string)
-	kind := jsonData["kind"].(string)
-	// 4. Construct the transparency log entry list
-	return []*protorekor.TransparencyLogEntry{
-		{
-			LogIndex: int64(logIndex),
-			LogId: &protocommon.LogId{
-				KeyId: logID,
-			},
-			KindVersion: &protorekor.KindVersion{
-				Kind:    kind,
-				Version: apiVersion,
-			},
-			IntegratedTime: int64(integratedTime),
-			InclusionPromise: &protorekor.InclusionPromise{
-				SignedEntryTimestamp: signedEntryTimestamp,
-			},
-			InclusionProof:    nil,
-			CanonicalizedBody: bodyBytes,
-		},
-	}, nil
-}
-
-// getVerificationMaterialX509CertificateChain returns the verification material X509 certificate chain from the simple signing layer
-func getVerificationMaterialX509CertificateChain(manifestLayer *v1.Descriptor) (*protobundle.VerificationMaterial_X509CertificateChain, error) {
-	// 1. Get the PEM certificate from the simple signing layer
-	pemCert := manifestLayer.Annotations["dev.sigstore.cosign/certificate"]
-	// 2. Construct the DER encoded version of the PEM certificate
-	block, _ := pem.Decode([]byte(pemCert))
-	if block == nil {
-		return nil, errors.New("failed to decode PEM block")
-	}
-	signingCert := protocommon.X509Certificate{
-		RawBytes: block.Bytes,
-	}
-	// 3. Construct the X509 certificate chain
-	return &protobundle.VerificationMaterial_X509CertificateChain{
-		X509CertificateChain: &protocommon.X509CertificateChain{
-			Certificates: []*protocommon.X509Certificate{&signingCert},
-		},
-	}, nil
-}
-
-// getBundleMsgSignature returns the bundle message signature from the simple signing layer
-func getBundleMsgSignature(simpleSigningLayer *v1.Descriptor) (*protobundle.Bundle_MessageSignature, error) {
-	// 1. Get the message digest algorithm
-	var msgHashAlg protocommon.HashAlgorithm
-	switch simpleSigningLayer.Digest.Algorithm {
-	case "sha256":
-		msgHashAlg = protocommon.HashAlgorithm_SHA2_256
-	default:
-		return nil, fmt.Errorf("unknown digest algorithm: %s", simpleSigningLayer.Digest.Algorithm)
-	}
-	// 2. Get the message digest
-	digest, err := hex.DecodeString(simpleSigningLayer.Digest.Hex)
-	if err != nil {
-		return nil, fmt.Errorf("error decoding digest: %w", err)
-	}
-	// 3. Get the signature
-	s := simpleSigningLayer.Annotations["dev.cosignproject.cosign/signature"]
-	sig, err := base64.StdEncoding.DecodeString(s)
-	if err != nil {
-		return nil, fmt.Errorf("error decoding manSig: %w", err)
-	}
-	// Construct the bundle message signature
-	return &protobundle.Bundle_MessageSignature{
-		MessageSignature: &protocommon.MessageSignature{
-			MessageDigest: &protocommon.HashOutput{
-				Algorithm: msgHashAlg,
-				Digest:    digest,
-			},
-			Signature: sig,
-		},
-	}, nil
 }
